Create utility fn for file checking
