+++ imagedephi/redact/build_redaction_plan.py
@@ -1,6 +1,7 @@
 from pathlib import Path
 
 from imagedephi.rules import FileFormat, Ruleset
+from imagedephi.utils.image import get_file_format_from_path
 
 from .dicom import DicomRedactionPlan
 from .redaction_plan import RedactionPlan
@@ -13,20 +14,6 @@ FILE_EXTENSION_MAP: dict[str, FileFormat] = {
     ".svs": FileFormat.SVS,
     ".dcm": FileFormat.DICOM,
 }
-
-
-def get_file_format_from_path(image_path: Path) -> FileFormat | None:
-    try:
-        data = open(image_path, "rb").read(132)
-    except PermissionError:
-        raise Exception(f"Could not open {image_path}")
-    else:
-        if data[:4] in (b"II\x2a\x00", b"MM\x00\x2a", b"II\x2b\x00", b"MM\x00\x2b"):
-            # tiff
-            return FileFormat.TIFF
-        elif data[128:] == b"DICM":
-            return FileFormat.DICOM
-    return None
 
 
 def build_redaction_plan(
